+++ functional/pipeline.py
@@ -20,6 +20,7 @@ from functional.util import (
     is_namedtuple,
     is_tabulatable,
     identity,
+    default_value,
 )
 from functional.io import WRITE_MODE, universal_write_open
 from functional import transformations
@@ -32,7 +33,9 @@ class Sequence(object):
     functional transformations and reductions in a data pipeline style
     """
 
-    def __init__(self, sequence, transform=None, engine=None, max_repr_items=None):
+    def __init__(
+        self, sequence, transform=None, engine=None, max_repr_items=None, no_wrap=None
+    ):
         # pylint: disable=protected-access
         """
         Takes a Sequence, list, tuple. or iterable sequence and wraps it around a Sequence object.
@@ -43,6 +46,7 @@ class Sequence(object):
         :param transform: transformation to apply
         :param engine: execution engine
         :param max_repr_items: maximum number of items to print with repr
+        :param no_wrap: default value of no_wrap for functions like first() or last()
         :return: sequence wrapped in a Sequence
         """
         self.engine = engine or ExecutionEngine()
@@ -58,6 +62,7 @@ class Sequence(object):
             raise TypeError("Given sequence must be an iterable value")
         if transform is not None:
             self._lineage.apply(transform)
+        self.no_wrap = no_wrap
 
     def __iter__(self):
         """
@@ -164,9 +169,9 @@ class Sequence(object):
         :return: concatenated sequence with other
         """
         if isinstance(other, Sequence):
-            return Sequence(self.sequence + other.sequence)
+            return Sequence(self.sequence + other.sequence, no_wrap=self.no_wrap)
         else:
-            return Sequence(self.sequence + other)
+            return Sequence(self.sequence + other, no_wrap=self.no_wrap)
 
     def _evaluate(self):
         """
@@ -185,9 +190,9 @@ class Sequence(object):
         sequence = None
         for transform in transforms:
             if sequence:
-                sequence = Sequence(sequence, transform=transform)
+                sequence = Sequence(sequence, transform=transform, no_wrap=self.no_wrap)
             else:
-                sequence = Sequence(self, transform=transform)
+                sequence = Sequence(self, transform=transform, no_wrap=self.no_wrap)
         return sequence
 
     @property
@@ -220,7 +225,7 @@ class Sequence(object):
             self._lineage = Lineage(engine=self.engine)
         return self
 
-    def head(self):
+    def head(self, no_wrap=None):
         """
         Returns the first element of the sequence.
 
@@ -234,11 +239,15 @@ class Sequence(object):
          ...
         IndexError: list index out of range
 
+        :param no_wrap: If set to True, the returned value will never be wrapped with Sequence
         :return: first element of sequence
         """
-        return _wrap(self.take(1)[0])
+        if default_value(no_wrap, self.no_wrap, False):
+            return self.sequence[0]
+        else:
+            return _wrap(self.take(1)[0])
 
-    def first(self):
+    def first(self, no_wrap=None):
         """
         Returns the first element of the sequence.
 
@@ -252,11 +261,12 @@ class Sequence(object):
          ...
         IndexError: list index out of range
 
+        :param no_wrap: If set to True, the returned value will never be wrapped with Sequence
         :return: first element of sequence
         """
-        return self.head()
+        return self.head(no_wrap=no_wrap)
 
-    def head_option(self):
+    def head_option(self, no_wrap=None):
         """
         Returns the first element of the sequence or None, if the sequence is empty.
 
@@ -266,13 +276,14 @@ class Sequence(object):
         >>> seq([]).head_option()
         None
 
+        :param no_wrap: If set to True, the returned value will never be wrapped with Sequence
         :return: first element of sequence or None if sequence is empty
         """
         if not self.sequence:
             return None
-        return self.head()
+        return self.head(no_wrap=no_wrap)
 
-    def last(self):
+    def last(self, no_wrap=None):
         """
         Returns the last element of the sequence.
 
@@ -286,11 +297,15 @@ class Sequence(object):
          ...
         IndexError: list index out of range
 
+        :param no_wrap: If set to True, the returned value will never be wrapped with Sequence
         :return: last element of sequence
         """
-        return _wrap(self.sequence[-1])
+        if default_value(no_wrap, self.no_wrap, False):
+            return self.sequence[-1]
+        else:
+            return _wrap(self.sequence[-1])
 
-    def last_option(self):
+    def last_option(self, no_wrap=None):
         """
         Returns the last element of the sequence or None, if the sequence is empty.
 
@@ -300,11 +315,12 @@ class Sequence(object):
         >>> seq([]).last_option()
         None
 
+        :param no_wrap: If set to True, the returned value will never be wrapped with Sequence
         :return: last element of sequence or None if sequence is empty
         """
         if not self.sequence:
             return None
-        return self.last()
+        return self.last(no_wrap=no_wrap)
 
     def init(self):
         """