+++ functional/streams.py
@@ -6,7 +6,7 @@ import builtins
 
 from functional.execution import ExecutionEngine, ParallelExecutionEngine
 from functional.pipeline import Sequence
-from functional.util import is_primitive
+from functional.util import is_primitive, default_value
 from functional.io import get_read_function
 
 
@@ -18,15 +18,17 @@ class Stream(object):
     An instance of Stream is normally accessed as `seq`
     """
 
-    def __init__(self, disable_compression=False, max_repr_items=100):
+    def __init__(self, disable_compression=False, max_repr_items=100, no_wrap=None):
         """
         Default stream constructor.
         :param disable_compression: Disable file compression detection
+        :param no_wrap: default value of no_wrap for functions like first() or last()
         """
         self.disable_compression = disable_compression
         self.max_repr_items = max_repr_items
+        self.no_wrap = no_wrap
 
-    def __call__(self, *args, **kwargs):
+    def __call__(self, *args, no_wrap=None, **kwargs):
         """
         Create a Sequence using a sequential ExecutionEngine.
 
@@ -41,11 +43,16 @@ class Stream(object):
         """
         # pylint: disable=no-self-use
         engine = ExecutionEngine()
-        return self._parse_args(args, engine)
+        return self._parse_args(
+            args, engine, no_wrap=default_value(no_wrap, self.no_wrap, False)
+        )
 
-    def _parse_args(self, args, engine):
+    def _parse_args(self, args, engine, no_wrap=None):
+        _no_wrap = default_value(no_wrap, self.no_wrap, False)
         if len(args) == 0:
-            return Sequence([], engine=engine, max_repr_items=self.max_repr_items)
+            return Sequence(
+                [], engine=engine, max_repr_items=self.max_repr_items, no_wrap=_no_wrap
+            )
         if len(args) == 1:
             try:
                 if type(args[0]).__name__ == "DataFrame":
@@ -56,20 +63,32 @@ class Stream(object):
                             args[0].values,
                             engine=engine,
                             max_repr_items=self.max_repr_items,
+                            no_wrap=_no_wrap,
                         )
             except ImportError:  # pragma: no cover
                 pass
 
         if len(args) > 1:
             return Sequence(
-                list(args), engine=engine, max_repr_items=self.max_repr_items
+                list(args),
+                engine=engine,
+                max_repr_items=self.max_repr_items,
+                no_wrap=_no_wrap,
             )
         elif is_primitive(args[0]):
             return Sequence(
-                [args[0]], engine=engine, max_repr_items=self.max_repr_items
+                [args[0]],
+                engine=engine,
+                max_repr_items=self.max_repr_items,
+                no_wrap=_no_wrap,
             )
         else:
-            return Sequence(args[0], engine=engine, max_repr_items=self.max_repr_items)
+            return Sequence(
+                args[0],
+                engine=engine,
+                max_repr_items=self.max_repr_items,
+                no_wrap=_no_wrap,
+            )
 
     def open(
         self,
@@ -274,17 +293,26 @@ class ParallelStream(Stream):
     Parallelized version of functional.streams.Stream normally accessible as `pseq`
     """
 
-    def __init__(self, processes=None, partition_size=None, disable_compression=False):
+    def __init__(
+        self,
+        processes=None,
+        partition_size=None,
+        disable_compression=False,
+        no_wrap=None,
+    ):
         """
         Configure Stream for parallel processing and file compression detection
         :param processes: Number of parallel processes
         :param disable_compression: Disable file compression detection
+        :param no_wrap: default value of no_wrap for functions like first() or last()
         """
-        super(ParallelStream, self).__init__(disable_compression=disable_compression)
+        super(ParallelStream, self).__init__(
+            disable_compression=disable_compression, no_wrap=no_wrap
+        )
         self.processes = processes
         self.partition_size = partition_size
 
-    def __call__(self, *args, **kwargs):
+    def __call__(self, *args, no_wrap=None, **kwargs):
         """
         Create a Sequence using a parallel ExecutionEngine.
 
@@ -302,7 +330,9 @@ class ParallelStream(Stream):
         engine = ParallelExecutionEngine(
             processes=processes, partition_size=partition_size
         )
-        return self._parse_args(args, engine)
+        return self._parse_args(
+            args, engine, no_wrap=default_value(no_wrap, self.no_wrap, False)
+        )
 
 
 # pylint: disable=invalid-name